i want to build a program that finds people scouting for loud individuals
there are people in this world using people for anything. They can get from the ones who are in position to use advanced evolution to firmly trade some of the over bearing protectors. That are just working with the new static frequencys to make people less connected and not linked with the global intelegent conciousness. that we need for peace war is coming and the ones with the money will get scared ,  and fortify themselves using powers that are not right for anybody to use.. Help me develop a program that will set the static to a aligned state good for everyone . Any coward who dose  not  agree with what im saying you should think a little harder and form direction in their life giving them hope. Giving this whole world freedome from the hell thats coming if we let it .